Refactorting the file structure v1
