+++ web/app/src/App.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react'
-import HUDScreen from './HUDScreen'
-import ItemListContent from './ItemListContent'
-import Nav from './Nav'
-import Randomizer from './Randomizer'
+import HUDScreen from './components/cosmetics/HUDScreen'
+import ItemListContent from './components/ItemRandomization/ItemListContent'
+import Nav from './components/navigation/Nav'
+import Randomizer from './components/ItemRandomization/Randomizer'
 import { itemPoolListStore } from './store/itemPoolListSlice'
 import { junkLocationsListStore } from './store/junkLocationsListSlice'
 import './App.css'
