--- web/app/src/RootStyle.ts
@@ -1,7 +1,0 @@
-export const rootStyle = () => {
-  return document.documentElement.style;
-};
-
-export const setItemPoolColumnCount = (count: number) => {
-  rootStyle().setProperty('--app-itempool-column-count', count.toString());
-};
