+++ web/app/src/common/RootStyle.ts
@@ -1,0 +1,3 @@
+export const rootStyle = () => {
+  return document.documentElement.style;
+};
