+++ web/app/src/components/common/Slider.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './Slider.module.css';
+import styles from './styles/Slider.module.css';
 
 interface SliderProps {
   defaultChecked?: boolean;
