+++ web/app/src/components/common/styles/Slider.module.css
