+++ web/app/src/index.tsx
@@ -4,7 +4,7 @@ import './index.css';
 import '@fontsource/inter/variable.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import store from './store';
+import store from './store/store';
 import { Provider } from 'react-redux';
 
 ReactDOM.render(
